requirements.txt
annotated-doc==0.0.4
fpdf2==2.8.7
pytest==9.0.2
annotated-types==0.7.0
anyio==4.13.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
confection==1.3.3
cryptography==46.0.6
cymem==2.0.13
distro==1.9.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.4
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
preshed==3.0.13
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pypdfium2==5.6.0
python-docx==1.2.0
requests==2.33.1
rich==14.3.3
setuptools==82.0.1
shellingham==1.5.4
smart_open==7.5.1
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
tenacity==9.1.4
thinc==8.3.13
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wasabi==1.1.3
weasel==1.0.0
websockets==16.0
wrapt==2.1.2
